Correct read all policy, add default users
Default read all policy did not have an 'all' resource attached so it didn't have an effect on users or groups, this has been corrected. Added a couple of default users to make testing easier and modified the user mapper, so user's emails are only revealed to a user with edit permissions on that user.