Setup dockerfiles and prod settings
Dockerfiles are in place for the first 4 services to deploy them in ECS. This uses the jar created by the 'bootJar' job. This commit also includes changes to how settings are included in production via environmental variables, how they're displayed in the settings endpoint, updates to the gradle version, and gradle build updates to allow the switch to java version 15.