Large user and access service rewrite
Main changes:

Refresh tokens have been introduced. Instead of access tokens being possible to directly refresh, there's a distinct new refresh token. A refresh token can only be used to refresh a session to get a new access/refresh token pair, and the much shorter lived access token is used for all other API requests.

Access tokens no longer contain any data related to groups a user is part of. This information is now only persisted within the access service, and is stored in a separate table instead of as part of a string in the session.

Misc changes:

* Removed `Assertions.` prefix from many test assertions in the access service.
* Switched access service integration tests to use test containers.
* Error codes are now references through constants, instead of directly as strings.
* DbLoader has been rearranged to make it easier to handle ephemeral test scripts.