Password reset for existing email registrations
The registration process has been tweaked so that registering with an existing email will cause a password reset email to be sent. The response returned to the user will be the same as if they'd registered properly, so no information on if the email was already registered or not will be divulged.